Add the stopParsingAtFirstUnknown parsing option
With this options, parsing arguments similar to docker is possible. For
example `docker run -i -t centos ls -l -a`, where the first parser
will parse out the command, `run` in this case, the second parser will
parse `-i` and `-t` and then pass the remaining arguments to another parser.
Before this change if the second parser had a `-l` or `-a` option they would
be parsed and not passed on.